Implemented separated inspection
