--- sonosthesia-audio-pipeline/process.py
@@ -1,8 +1,0 @@
-from collections import namedtuple
-
-# Trying to pull out some common behaviour between sonosthesia-pipeline and preview but ends up being more of a mess
-
-AnalysisData = namedtuple('AnalysisData', ['time', 'rms', 'lows', 'mids', 'highs', 'centroid', 'onset'])
-
-def process_audio(y, sr):
-    pass
